Organized Files, Added Mock Restaurants, Added SplashScreen, Restructured Code, Preloaded Assets
I added several mock restaurants in a flatlist on the homepage to represent the recent trips. I also moved all assets to separate folders or created specific subfolders for each asset. To prevent any assets to load while the user is looking at the screen, I implemented a splash screen which will show to the user while all of the assets(images,fonts, user information is preloading). Finally, I restructured the code in most of the files to prevent recurring bugs.